Adding diagnostic scripts
These scripts can be used to debug the camera
system if anything is not working. json_parser
uses json files collected and provides various
plots and metrics. check_camera_settings shows
the values for all parameters for each camera on
the network and outputs a CSV with all params and
a CSV with params that differ between cameras.