GUI: allow swapping scenes to annotate
The main logic is now working. Now the data is processing much faster
I also dropped the socket based streaming. Can always bring it back if
we really need it, but the animation locked up a lot during replay so
it wasn't exactly like a transparent background buffering.

Need to reanimate the data too, and then the annotation button should
work.