Docker: Can record from one camera
This code needs to be cleaned up from various hacks.
Changes in this commit reordering Dockerfile, adding
a local datajoint_config (not committed) to replace the
default one from the posepipe image. Adding the config
file (not committed) into the container. Updating the
bash script run at the end of the Dockerfile to run both
the frontend and backend. Updating the path that the
backend looks in for config files. Updating base URL
for frontend. Updating allowed origins in fastapi.py