Docker: Adding support for 60 Hz recording
The recording api now expects an input config file to
have both exposure_time and frame_rate set. For frame_rate
if it is set to 60, there will be a binning of 2 applied
vertically and horizontally, otherwise no binning. The
exposure_time from the config file is now passed to
the init_camera() method.