Docker: Fixing bug switching between 30Hz and 60Hz
To swap between 30 and 60Hz, we implement a binning in both
directions. One problem was that the binning change to the
frame size would be persistent unless we restarted the cameras.
Many of the settings are not writable or have to be within the
range of the max height and width which is defined by binning.
The workaround for this is to always set binning = 1 when we
initialize cameras.