[Acquisition] Removing the need for a config file
The way the code was written, we always had to provide
a config file to run. Now, a number of cameras can be
specified and that number will be randomly selected
from the cameras available on the interface. If there
is no config, default values for exposure time and
frame rate are used.