[Docker] Updating Dockerfile for calibration
Adding installation for tbb and numba to the
Dockerfile to attempt to fix the calibration
code not using parallelization. Currently,
the tbb version is too old so parallelization
is disabled. This commit does not fix the
issue and can potentially be removed in the
future after the root cause is determined.