Issue #24: Updating recording api to support continuous recording
Updates include always setting line selector and line source: Line1/ExposureActive
for continuous and Line0/Off for max frame recording. Updating a bug in the code
that creates the new video file to write to.

There is still an outstanding issue where max frame recording records in chunks. Just
need to add a condition to check for 'continuous'. Also need to confirm filenames are
written correctly to the database and update GUI to show recording progress. Finally,
need to check how many frames are lost when starting the new file and clean up the code.