Issue #24: Adding trigger back into recording api
When adding continuous recording mode, we originally removed
the triggering at the start of the recordings. I am adding this
back in because it is needed for the cameras to start recording
at the same time. So the only changes that have been made to
allow continuous recording is to create new video files every
n frames and to have the exposure pulses readable from the
GPIO pin.