arduino_sync: Receiving trigger for every frame and frame count from serial data
The code has been updated to get triggered on each frame by a signal
from the arduino. This is hardcoded currently and will need to be
updated. Additionally, we are able to receive the serial data from the
arduino, read the chunk serial data, and reconstruct the frame count
sent by the arduino.