camera_gpio: Better integrating GPIO development
Previously, toggling different configurations for
the GPIO pins was not possible from the config file
as everything was hardcoded. This commit cleans up
some of the code and exposes some of the variables
to the config file so the GPIO pins can be
controlled independently. This still needs
additional cleanup