continuous_dev: Updating metadata writing method
Updating the write_metadata_queue method to calculate timespread
before writing, and properly handling the queue so it stopped
cleanly when the stop button was pressed. Also updating the
write_image_queue method to release the previous video before
writing to a new one. Finally, removing old json building
code. Still to fix is passing the records to the backend properly.