continuous_dev: Updating the JSON to match old format
The fields in the JSON file now line up again with the old method
of writing the JSONs. There are some additional fields added that
will likely be removed after some additional debugging. One issue
is that one frame is missing from the JSONs when a new video
segment is started. This will be corrected in the next commit.