env_file: Using docker compose rather than docker-compose
docker-compose is no longer supported by Docker in favor of
the newer docker compose. Also updating some of the names
of the environment variables that were added to be more consistent
with the rest of the system. Additionally, just adding the timezone
mount hardcoded in the docker-compose file. Finally, using
"docker compose run mocap" rather than "docker compose up". This
is mainly aesthetic as docker-compose doesn't work well with tqdm.
This can be changed in the future if we add more services.