reading_threads: Fixing silent bug and pytest
There was a bug in the frame monitoring code which
did not throw an error when it occurred. Separately,
there was an issue with how one of the pytests was
writted so the error was not discovered. This commit
fixes those and does some additional cleanup. This silent
bug erroneously led me to believe that the efficiency
issues were resolved by removing tqdm. It was a step in
the right direction but there are still fixes needed.