acquisition: fixing bug with stop button logic
The stop button logic in the process_synchronized_metadata
caused a silent exception which failed successfully. I have
updated this to explicitly break out of the loop after
finishing the cleanup frames. Also cleaning up unused code
and refactoring some of the code based on comments from
the pull request.