continuous: Updating code to work for continuous mode again
This is still in progress. So far I have fixed bugs where
each video segment was not written as a separate record, fixed
the progress update in the GUI, and updated the logic when new
continuous trials are started to skip frame 0. Some outstanding
issues are the preview does not work properly and there is an
extra json file written with no videos when the stop button is
pressed.