continuous: Trying to fix file naming bug
This commit is working on fixing the issue where the
video filenames do not update properly causing the previous
video segment to always be overwritten resulting in one less
set of videos. The current approach does not work 100% correctly,
we end up with the correct number of videos but some of the videos
are not saved properly and the json file has varying numbers of
frames saved. Will update in future commit.