continuous: Passing filename to video and json methods
Working on the issue from the previous commit. Now getting the
file name that videos/json should be written to since it is
explicitly passed in the respective queues. Also, consolidating
the video writing method. Still having an issue where filenames
are changed by each camera thread so some filenames are skipped
so fixing that now.